Start implementation of PeopleSoft message handling
When PeopleSoft POSTs a rowset-based message to the PSSyncCollector
resource, being handling it by parsing the XML for the PeopleSoft
message name and the field types.

The field_type function is passsed to element_to_obj() so that the type
attribute is used as the value for the resulting child elements. This
produces key-value pairs of fieldtype name and fieldtype type for later
use.