If key_attribute is not defined for record_type, optionally guess
In key_for_record, if no key_attribute is defined in
key_attributes_by_record_name and guess is True, find the root
record_data attribute whose key is the most similar to the record_type,
and use that as the key_attribute. This provides a reasonable guess for
the key_attribute in limited sample cases, so far.